--- components/AppLogo.vue
@@ -1,102 +1,0 @@
-<template>
-  <svg
-    version="1.1"
-    x="0px"
-    y="0px"
-    viewBox="0 0 80.781 88.847997"
-    xml:space="preserve"
-    id="svg3"
-    sodipodi:docname="noun-ace-card-900153.svg"
-    width="32.6247984"
-    height="39.0784"
-    inkscape:version="1.3.1 (91b66b0783, 2023-11-16)"
-    inkscape:export-filename="noun-ace-card-900153.png"
-    inkscape:export-xdpi="608.4599"
-    inkscape:export-ydpi="608.4599"
-    xmlns:inkscape="http://www.inkscape.org/namespaces/inkscape"
-    xmlns:sodipodi="http://sodipodi.sourceforge.net/DTD/sodipodi-0.dtd"
-    xmlns:xlink="http://www.w3.org/1999/xlink"
-    xmlns="http://www.w3.org/2000/svg"
-    xmlns:svg="http://www.w3.org/2000/svg"
-  >
-    <defs id="defs3">
-      <linearGradient id="linearGradient3" inkscape:collect="always">
-        <stop
-          style="stop-color: #005c99; stop-opacity: 1"
-          offset="0"
-          id="stop3"
-        />
-        <stop
-          style="stop-color: #009dcc; stop-opacity: 1"
-          offset="1"
-          id="stop4"
-        />
-      </linearGradient>
-      <linearGradient
-        inkscape:collect="always"
-        xlink:href="#linearGradient3"
-        id="linearGradient4"
-        x1="9.651989"
-        y1="49.975994"
-        x2="90.433334"
-        y2="49.975994"
-        gradientUnits="userSpaceOnUse"
-      />
-      <linearGradient
-        inkscape:collect="always"
-        xlink:href="#linearGradient3"
-        id="linearGradient1"
-        gradientUnits="userSpaceOnUse"
-        x1="9.651989"
-        y1="49.975994"
-        x2="90.433334"
-        y2="49.975994"
-      />
-      <linearGradient
-        inkscape:collect="always"
-        xlink:href="#linearGradient3"
-        id="linearGradient2"
-        gradientUnits="userSpaceOnUse"
-        x1="9.651989"
-        y1="49.975994"
-        x2="90.433334"
-        y2="49.975994"
-      />
-    </defs>
-    <sodipodi:namedview
-      id="namedview3"
-      pagecolor="#ffffff"
-      bordercolor="#000000"
-      borderopacity="0.25"
-      inkscape:showpageshadow="2"
-      inkscape:pageopacity="0.0"
-      inkscape:pagecheckerboard="0"
-      inkscape:deskcolor="#d1d1d1"
-      inkscape:zoom="4.1556837"
-      inkscape:cx="15.520912"
-      inkscape:cy="88.91437"
-      inkscape:window-width="1912"
-      inkscape:window-height="1032"
-      inkscape:window-x="0"
-      inkscape:window-y="0"
-      inkscape:window-maximized="1"
-      inkscape:current-layer="svg3"
-    />
-    <g
-      id="g2"
-      transform="translate(-9.6519887,-5.5519887)"
-      style="fill: url(#linearGradient4)"
-    >
-      <path
-        d="m 50.5,11.2 c 3.2,-0.6 6.3,0.2 9,2 2.6,1.8 4.4,4.6 4.9,7.7 l 10,56.5 c 0.5,3 -0.1,6 -1.8,8.6 3.8,-0.5 7,-3.4 7.7,-7.4 l 10,-56.5 c 0.4,-2.4 -0.1,-4.7 -1.5,-6.7 -1.4,-2 -3.4,-3.3 -5.8,-3.7 L 49,5.7 C 44.1,4.8 39.4,8.1 38.6,13 v 0.3 z"
-        id="path1"
-        style="fill: url(#linearGradient1)"
-      />
-      <path
-        d="m 19.7,87 c 0.8,4.3 4.5,7.4 8.9,7.4 0.5,0 1,0 1.6,-0.1 l 33.9,-6 c 2.4,-0.4 4.4,-1.7 5.8,-3.7 1.4,-2 1.9,-4.4 1.5,-6.7 L 61.4,21.4 C 60.7,17.1 57,14 52.6,14 v 0 c -0.5,0 -1,0 -1.6,0.1 l -33.9,6 C 12.2,21 8.9,25.6 9.8,30.5 Z M 38.1,40 c 0,0 15.4,6.5 16.9,15 0.7,3.8 -2.3,7.6 -6.6,8.4 -1.8,0 -3.4,-0.6 -4.7,-1.7 1.4,2.2 3.3,4.2 5.4,5.7 l -11.8,2.1 c 1.5,-2.2 2.5,-4.7 3.1,-7.2 -0.8,1.5 -2.2,2.6 -3.8,3.2 -4.4,0.8 -8.4,-1.7 -9.1,-5.6 C 25.9,51.3 38.1,40 38.1,40 Z"
-        id="path2"
-        style="fill: url(#linearGradient2)"
-      />
-    </g>
-  </svg>
-</template>
