+++ components/WIP.vue
@@ -3,7 +3,7 @@
     <div class="modal-box bg-slate-950">
       <div class="w-full justify-center text-center">
         <h2 class="font-bold text-2xl capitalize mb-5">Currently Work in Progress...</h2>
-        <p>In the mean time you can check out the <a href="/faucet" class="link">faucet</a> to get test tokens as well as the <a href="/documentation" class="link">docs</a> to learn more about the project.</p>
+        <p>In the mean time you can check out the <nuxt-link href="/faucet" class="link">faucet</nuxt-link> to get test tokens as well as the <nuxt-link href="/documentation" class="link">docs</nuxt-link> to learn more about the project.</p>
       </div>
     </div>
     <form method="dialog" class="modal-backdrop">
