+++ pages/features/index.vue
@@ -1,0 +1,17 @@
+<template>
+  <div class="w-full min-h-[67vh] flex items-center justify-center">
+    <Head> </Head>
+    <title>BottleHub - Gambling on The Edge</title>
+    <div class="hero-content max-w-full w-full"></div>
+  </div>
+</template>
+
+<script lang="js" setup>
+definePageMeta({
+  pageTransition: {
+    name: "slide-right",
+    mode: "out-in",
+  }
+});
+wip.showModal();
+</script>
