--- pages/lobby.vue
@@ -1,23 +1,0 @@
-<template>
-  <Head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="preconnect" href="https://fonts.googleapis.com" />
-    <link rel="preconnect" href="https://fonts.gstatic.com" />
-    <link
-      href="https://fonts.googleapis.com/css2?family=Libre+Baskerville:ital,wght@0,700;1,400&display=swap"
-      rel="stylesheet"
-    />
-  </Head>
-  <title>BottleHub - Gambling on The Edge</title>
-  <div class="h-screen flex">
-    <div class="w-24 flex justify-center items-center">
-      <AppNavigation />
-    </div>
-    <div class="w-full flex justify-center items-center">
-    <AppHub />
-    </div>
-  </div>
-</template>
-<script setup lang="ts"></script>
-<style></style>
